mutators qui marchent

+++ src/main/java/fr/istic/vv/Mutators.java
@@ -3,8 +3,10 @@ package fr.istic.vv;
 import javassist.*;
 import javassist.expr.ExprEditor;
 import javassist.expr.FieldAccess;
+import javassist.expr.MethodCall;
 
 import java.io.File;
+import java.util.Set;
 
 public class Mutators {
 
@@ -31,11 +33,41 @@ public class Mutators {
         return ctClass;
     }
 
-    public static void revertAllChanges(){
+    public static CtClass setBooleanMethodsTo(CtClass ctClass, final boolean bool){
+        if(ctClass.isInterface()){
+            return ctClass;
+        }
+        for(CtMethod ctMethod : ctClass.getDeclaredMethods()){
+            try {
+                if(ctMethod.getReturnType().getName().equals("boolean")){
+                    ctMethod.setBody("return " + bool + ";");
+                    /*ctMethod.instrument(new ExprEditor() {
+                        @Override
+                        public void edit(MethodCall mc) throws CannotCompileException {
+                            mc.replace("{ $_ = true; }");
+                        }
+                    });*/
+                }
+            } catch (NotFoundException e) {
+                e.printStackTrace();
+            } catch (NullPointerException e){
+                e.printStackTrace();
+            } catch (CannotCompileException e) {
+                e.printStackTrace();
+            }
+        }
+        return ctClass;
+    }
+
+    public static void deleteTargetClasses(Set<CtClass> toDelete){
         File target = new File("target/classes/fr/istic/vv");
         for(String file : target.list()){
-            File currentFile = new File(target.getPath(),file);
-            currentFile.delete();
+            for(CtClass ctClass : toDelete){
+                if(file.contains(ctClass.getSimpleName())){
+                    File currentFile = new File(target.getPath(),file);
+                    currentFile.delete();
+                }
+            }
         }
     }
 }
